updated the return type of the function

+++ iceoryx_eth/include/iceoryx_eth/eth/eth_reader.hpp
@@ -41,7 +41,7 @@ class ethDataReader : public DataReader
     void connect() noexcept override;
 
     uint8_t setUniqueCode(const iox::capro::ServiceDescription&);
-    uint8_t getUniqueCode();
+    uint64_t getUniqueCode();
 
     iox::cxx::optional<uint64_t> peekNextSize() override;
